ongoing work

+++ model/src/main/java/com/serveeasy/model/DomainObject.java
@@ -3,6 +3,8 @@ package com.serveeasy.model;
 import java.io.Serializable;
 
 /**
+ * Abstract type for all domain objects keeping the synthetic pk as long value
+ * <p/>
  * User: cristian.popovici
  */
 public abstract class DomainObject implements Serializable {
@@ -13,7 +15,7 @@ public abstract class DomainObject implements Serializable {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
